Extract description tooltip
